requirements.txt
tensorflow
numpy
beautifulsoup4
requests
Pillow
matplotlib
